refactor: move worker, postgres, sqlite under experimental/
Nest unstable submodules to signal API is not yet finalized:
- worker/ → experimental/worker/
- postgres/ → experimental/store/postgres/
- sqlite/ → experimental/store/sqlite/

Postgres and SQLite grouped under store/ as sibling persistence
implementations. All module paths, imports, replace directives,
and documentation updated.
