+++ README.md
@@ -126,9 +126,11 @@ For a turnkey queue-backed runner and a Postgres-backed
 implementation of every persistence interface the engine needs, see
 [`docs/worker.md`](docs/worker.md) and
 [`docs/postgres.md`](docs/postgres.md). The
-[`worker/`](worker/) and [`postgres/`](postgres/) submodules ship
-with this repo but live in their own Go modules so the root module
-stays stdlib-only.
+[`experimental/worker/`](experimental/worker/) and
+[`experimental/store/postgres/`](experimental/store/postgres/)
+submodules ship with this repo but live in their own Go modules so
+the root module stays stdlib-only. Everything under `experimental/`
+is functional but the API is not yet stable.
 
 ## Reference
