+++ experimental/store/postgres/credits.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/deepnoodle-ai/workflow/worker"
+	"github.com/deepnoodle-ai/workflow/experimental/worker"
 )
 
 // Debit implements worker.CreditStore. Idempotent per (run_id, "debit").
