+++ experimental/store/postgres/doc.go
@@ -1,7 +1,7 @@
 // Package postgres provides a Postgres-backed Store that implements
 // the four interfaces the workflow engine and its worker need:
 //
-//   - [github.com/deepnoodle-ai/workflow/worker.QueueStore] — the run
+//   - [github.com/deepnoodle-ai/workflow/experimental/worker.QueueStore] — the run
 //     queue, claim, heartbeat, reaper, and terminal state.
 //   - [github.com/deepnoodle-ai/workflow.Checkpointer] — via
 //     [Store.NewCheckpointer], lease-fenced checkpoint persistence
@@ -19,5 +19,5 @@
 // All writes to workflow_runs that belong to a running execution
 // fence on (claimed_by, attempt) so that a worker that has lost its
 // lease cannot corrupt a newer attempt's state. Fencing failures
-// surface as [github.com/deepnoodle-ai/workflow/worker.ErrLeaseLost].
+// surface as [github.com/deepnoodle-ai/workflow/experimental/worker.ErrLeaseLost].
 package postgres
