+++ experimental/store/sqlite/events.go
@@ -1,14 +1,13 @@
-package postgres
+package sqlite
 
 import (
 	"context"
+	"database/sql"
 	"encoding/json"
 	"fmt"
 	"time"
 
-	"github.com/jackc/pgx/v5"
-
-	"github.com/deepnoodle-ai/workflow/worker"
+	"github.com/deepnoodle-ai/workflow/experimental/worker"
 )
 
 // AppendEvent implements worker.EventStore.
@@ -17,14 +16,14 @@ func (s *Store) AppendEvent(ctx context.Context, event *worker.Event) error {
 	if event.Payload != nil {
 		b, err := json.Marshal(event.Payload)
 		if err != nil {
-			return fmt.Errorf("postgres: marshal event payload: %w", err)
+			return fmt.Errorf("sqlite: marshal event payload: %w", err)
 		}
 		payload = b
 	}
-	err := s.pool.QueryRow(ctx, `
+	err := s.db.QueryRowContext(ctx, `
 		INSERT INTO workflow_events (
 			run_id, event_type, attempt, worker_id, step_name, payload, created_at
-		) VALUES ($1,$2,$3,$4,$5,$6,$7)
+		) VALUES (?,?,?,?,?,?,?)
 		RETURNING seq
 	`,
 		event.RunID,
@@ -33,55 +32,55 @@ func (s *Store) AppendEvent(ctx context.Context, event *worker.Event) error {
 		event.WorkerID,
 		event.StepName,
 		payload,
-		event.CreatedAt,
+		formatTime(event.CreatedAt),
 	).Scan(&event.Seq)
 	if err != nil {
-		return fmt.Errorf("postgres: append event: %w", err)
+		return fmt.Errorf("sqlite: append event: %w", err)
 	}
 	return nil
 }
 
 // ListEvents implements worker.EventStore.
 func (s *Store) ListEvents(ctx context.Context, runID string, afterSeq int64) ([]*worker.Event, error) {
-	rows, err := s.pool.Query(ctx, `
+	rows, err := s.db.QueryContext(ctx, `
 		SELECT seq, run_id, event_type, attempt, worker_id, step_name, payload, created_at
 		FROM workflow_events
-		WHERE run_id = $1 AND seq > $2
+		WHERE run_id = ? AND seq > ?
 		ORDER BY seq ASC
 	`, runID, afterSeq)
 	if err != nil {
-		return nil, fmt.Errorf("postgres: list events: %w", err)
+		return nil, fmt.Errorf("sqlite: list events: %w", err)
 	}
 	defer rows.Close()
 
 	var out []*worker.Event
 	for rows.Next() {
 		var (
-			e       worker.Event
-			payload []byte
+			e         worker.Event
+			payload   sql.NullString
+			createdAt string
 		)
 		if err := rows.Scan(&e.Seq, &e.RunID, &e.EventType, &e.Attempt,
-			&e.WorkerID, &e.StepName, &payload, &e.CreatedAt); err != nil {
-			return nil, fmt.Errorf("postgres: scan event: %w", err)
+			&e.WorkerID, &e.StepName, &payload, &createdAt); err != nil {
+			return nil, fmt.Errorf("sqlite: scan event: %w", err)
 		}
-		if len(payload) > 0 {
-			_ = json.Unmarshal(payload, &e.Payload)
+		e.CreatedAt = parseTime(createdAt)
+		if payload.Valid && payload.String != "" {
+			_ = json.Unmarshal([]byte(payload.String), &e.Payload)
 		}
 		out = append(out, &e)
 	}
-	if err := rows.Err(); err != nil && err != pgx.ErrNoRows {
-		return nil, err
-	}
-	return out, nil
+	return out, rows.Err()
 }
 
 // CleanupEvents implements worker.EventStore.
 func (s *Store) CleanupEvents(ctx context.Context, olderThan time.Time) (int, error) {
-	tag, err := s.pool.Exec(ctx, `
-		DELETE FROM workflow_events WHERE created_at < $1
-	`, olderThan)
+	result, err := s.db.ExecContext(ctx, `
+		DELETE FROM workflow_events WHERE created_at < ?
+	`, olderThan.UTC().Format(timeFormat))
 	if err != nil {
-		return 0, fmt.Errorf("postgres: cleanup events: %w", err)
+		return 0, fmt.Errorf("sqlite: cleanup events: %w", err)
 	}
-	return int(tag.RowsAffected()), nil
+	n, _ := result.RowsAffected()
+	return int(n), nil
 }
