+++ experimental/worker/go.mod
@@ -1,0 +1,5 @@
+module github.com/deepnoodle-ai/workflow/experimental/worker
+
+go 1.26.1
+
+replace github.com/deepnoodle-ai/workflow => ../../
