--- worker/go.mod
@@ -1,5 +1,0 @@
-module github.com/deepnoodle-ai/workflow/worker
-
-go 1.26.1
-
-replace github.com/deepnoodle-ai/workflow => ../
